(e2e) chore: split up coverageReport test
